### 2017-09-05 #392 @whedon
[PRE REVIEW]: ECNet: Large scale machine learning projects for fuel property prediction
**Submitting author:** @TJKessler (<a href="http://orcid.org/0000-0002-7363-4050">Travis Kessler</a>)
**Repository:** <a href="https://github.com/TJKessler/ECNet" target ="_blank">https://github.com/TJKessler/ECNet</a>
**Version:** v1.2.4.dev1
