### 2017-09-11 #392 @TJKessler
@arokem - @krother or @jkahn from the provided list of JOSS reviewers would be a good fit
