### 2017-09-11 #392 @arokem
:wave: @krother : would you be willing to review this submission for JOSS?
