### 2017-09-12 #401 @whedon
[REVIEW]: ECNet: Large scale machine learning projects for fuel property prediction
**Submitting author:** @TJKessler (<a href="http://orcid.org/0000-0002-7363-4050">Travis Kessler</a>)
**Repository:** <a href="https://github.com/TJKessler/ECNet" target ="_blank">https://github.com/TJKessler/ECNet</a>
**Version:** v1.2.4.dev1

- [x] **Version:** Does the release version given match the GitHub release (v1.2.4.dev1)?
