### 2017-09-12 #392 @arokem
Thanks @krother! I am closing this now - head over to https://github.com/openjournals/joss-reviews/issues/401 for the review issue, including explanations and a check list for you to go through with the criteria for the review.
