### 2017-09-13 7a95aa7
added support for saving multiple outputs (per build) to output file
