### 2017-09-13 b8d5d51
added support for softmax layers
