### 2017-09-13 17c506a
fixed issue with select_best method where a final net could sometimes not be selected
