+++ ecnet/server.py
@@ -72,13 +72,10 @@ class Server:
 	def fit_mlp_model(self, *args):
         ### PROJECT ###
 		if self.folder_structs_built == True:
-			self.fit_rmse_list = []
 			for build in range(0,self.vars['project_num_builds']):
-				build_rmse_list = []
 				if self.vars['project_print_feedback'] == True:
 					print("Build %d of %d"%(build+1,self.vars['project_num_builds']))
 				for node in range(0,self.vars['project_num_nodes']):
-					node_rmse_list = []
 					if self.vars['project_print_feedback'] == True:
 						print("Node %d of %d"%(node+1,self.vars['project_num_nodes']))
 					for trial in range(0,self.vars['project_num_trials']):
@@ -93,8 +90,6 @@ class Server:
 							sys.exit()
 						try:
 							res = self.model.test_new(self.data.x)
-							rmse = ecnet.error_utils.calc_rmse(res, self.data.y)
-							node_rmse_list.append(rmse)
 						except:
 							raise
 							sys.exit()
@@ -107,8 +102,6 @@ class Server:
 							self.data.shuffle('l', 'v', data_split = self.vars['data_split'])
 						elif 'shuffle_lvt' in args:
 							self.data.shuffle('l', 'v', 't', data_split = self.vars['data_split'])
-					build_rmse_list.append(node_rmse_list)
-				self.fit_rmse_list.append(build_rmse_list)
 		### SINGLE NET ###
 		else:
 			self.model = create_model(self)
@@ -127,13 +120,10 @@ class Server:
 	def fit_mlp_model_validation(self, *args):
 		### PROJECT ###
 		if self.folder_structs_built == True:
-			self.fit_rmse_list = []
 			for build in range(0,self.vars['project_num_builds']):
-				build_rmse_list = []
 				if self.vars['project_print_feedback'] == True:
 					print("Build %d of %d"%(build+1,self.vars['project_num_builds']))
 				for node in range(0,self.vars['project_num_nodes']):
-					node_rmse_list = []
 					if self.vars['project_print_feedback'] == True:
 						print("Node %d of %d"%(node+1,self.vars['project_num_nodes']))
 					for trial in range(0,self.vars['project_num_trials']):
@@ -148,8 +138,6 @@ class Server:
 							sys.exit()
 						try:
 							res = self.model.test_new(self.data.x)
-							rmse = ecnet.error_utils.calc_rmse(res, self.data.y)
-							node_rmse_list.append(rmse)
 						except:
 							raise
 							sys.exit()
@@ -162,8 +150,6 @@ class Server:
 							self.data.shuffle('l', 'v', data_split = self.vars['data_split'])
 						elif 'shuffle_lvt' in args:
 							self.data.shuffle('l', 'v', 't', data_split = self.vars['data_split'])
-					build_rmse_list.append(node_rmse_list)
-				self.fit_rmse_list.append(build_rmse_list)
 		### SINGLE NET ###
 		else:
 			self.model = create_model(self)
@@ -186,36 +172,23 @@ class Server:
 			sys.exit()
 		### PROJECT ###
 		else:
-			# if you fit the model in the same script (self.fit_rmse_list is made)
-			try:
-				for build in range(len(self.fit_rmse_list)):
-					for node in range(len(self.fit_rmse_list[build])):
-						min_idx = 0
-						for error in range(len(self.fit_rmse_list[build][node])):
-							if self.fit_rmse_list[build][node][error] < self.fit_rmse_list[build][node][min_idx]:
-								min_idx = error
[…]
