### 2017-09-13 6a8f866
Update paper.md
