### 2017-09-13 #401 @krother
Thanks for the quick and precise repsonse. All issues resolved. Good job!
