### 2017-09-15 1d43f0f
removed redundant code for determining error function; removed unnecessary exception testing
