+++ ecnet/server.py
@@ -35,26 +35,14 @@ class Server:
 	def import_data(self, data_filename = None):
 		if data_filename is None:
 			data_filename = self.vars['data_filename']
-		try:
-			self.data = ecnet.data_utils.initialize_data(data_filename)
-		except:
-			raise
-			sys.exit()
-		try:
-			self.data.build()
-		except:
-			raise
-			sys.exit()
-		try:
-			self.data.buildTVL(self.vars['data_sort_type'], self.vars['data_split'])
-			if self.vars['normals_use'] == True:
-				self.data.normalize('./tmp/normal_params')
-			self.data.applyTVL()
-			self.data.package()
-		except:
-			raise
-			sys.exit()
-	
+		self.data = ecnet.data_utils.initialize_data(data_filename)
+		self.data.build()
+		self.data.buildTVL(self.vars['data_sort_type'], self.vars['data_split'])
+		if self.vars['normals_use'] == True:
+			self.data.normalize('./tmp/normal_params')
+		self.data.applyTVL()
+		self.data.package()
+
 	### Determines which 'param_num' parameters contribute to an accurate output; supply the number of parameters to limit to, and the output filename
 	def limit_parameters(self, param_num, limited_database_output_filename):
 		params = ecnet.limit_parameters.limit(self, param_num)
@@ -62,12 +50,8 @@ class Server:
 
 	### Creates the save environment
 	def create_save_env(self):
-		try:
-			create_folder_structure(self)
-		except:
-			raise
-			sys.exit()
-	
+		create_folder_structure(self)
+
 	### Fits the model(s) using predetermined number of learning epochs	
 	def fit_mlp_model(self, *args):
         ### PROJECT ###
@@ -83,21 +67,9 @@ class Server:
 							print("Trial %d of %d"%(trial+1,self.vars['project_num_trials']))
 						self.output_filepath = os.path.join(os.path.join(os.path.join(self.vars['project_name'], self.build_dirs[build]), self.node_dirs[build][node]), "model_output" + "_%d"%(trial + 1))
 						self.model = create_model(self)
-						try:
-							self.model.fit(self.data.learn_x, self.data.learn_y, self.vars['learning_rate'], self.vars['train_epochs'])
-						except:
-							raise
-							sys.exit()
-						try:
-							res = self.model.test_new(self.data.x)
-						except:
-							raise
-							sys.exit()
-						try:
-							self.model.save_net(self.output_filepath)
-						except:
-							raise
-							sys.exit()
+						self.model.fit(self.data.learn_x, self.data.learn_y, self.vars['learning_rate'], self.vars['train_epochs'])
+						res = self.model.test_new(self.data.x)
+						self.model.save_net(self.output_filepath)
 						if 'shuffle_lv' in args:
 							self.data.shuffle('l', 'v', data_split = self.vars['data_split'])
 						elif 'shuffle_lvt' in args:
@@ -105,16 +77,8 @@ class Server:
 		### SINGLE NET ###
 		else:
 			self.model = create_model(self)
-			try:
-				self.model.fit(self.data.learn_x, self.data.learn_y, self.vars['learning_rate'], self.vars['train_epochs'])
-			except:
-				raise
-				sys.exit()
-			try:
-				self.model.save_net("./tmp/model_output")
-			except:
-				raise
-				sys.exit()
+			self.model.fit(self.data.learn_x, self.data.learn_y, self.vars['learning_rate'], self.vars['train_epochs'])
+			self.model.save_net("./tmp/model_output")
 
 	### Fits the model(s) using validation RMSE cutoff method, or max epochs
[…]
