### 2017-09-15 40d0f5a
removed redundant error determination and exception testing code
