### 2017-09-15 a5ff453
added dataset arguments for select_best()
