### 2017-09-15 0645530
fixed select_best() syntax
