### 2017-09-15 2a8474d
cleaned up code
