### 2017-09-18 #401 @arokem
Thank you both for the quick turnaround here. @TJKessler : Your next step is to make an archive from the current version (e.g., using Zenodo) and to post the DOI of the archive here.
