### 2017-09-18 7db2cee
Update paper.md
