### 2017-09-18 #401 @arfon
@TJKessler - Could you move the references you currently have in the `paper.md` file into a `paper.bib` file and cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
