### 2017-09-18 df70c89
updated references, included all found in paper.md
