+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@online{ECNet (GitHub),
+@online{
   author = {Travis Kessler}
   title = {ECNet: Large scale machine learning projects for fuel property prediction}
   year = 2017
@@ -6,15 +6,15 @@
   urldate = {2017-10-04}
 }
 
-@online{ECNet (PyPI),
+@online{
   author = {Travis Kessler}
-  title = {ECNet 1.2.4.dev1}
+  title = {ECNet 1.2.5.dev1}
   year = 2017
   url = {https://pypi.python.org/pypi/ecnet/}
   urldate = {2017-10-04}
 }
 
-@online{UML ECRL,
+@online{
   author = {John Hunter Mack}
   title = {Energy and Combustion Research Laboratory}
   year = 2017
@@ -22,9 +22,32 @@
   urldate = {2017-10-04}
 }
 
-@misc{
-  author = {Travis Kessler}
+@PROCEEDINGS{
+  author = {T. Sennott, C. Gotianun, R. Serres, M. Ziabasharhagh, J. H. Mack, R. Dibble}
+  title = {Artificial Neural Network for Predicting Cetane Number of Biofuel Candidates Based on Molecular Structure}
+  booktitle = {ASME 2013 Internal Combustion Engine Division Fall Technical Conference}
+  publisher = {ASME}
+  address = {Dearborn, Michigan}
+  year = {2013}
+  doi = {10.1115/ICEF2013-19185}
+}
+
+@PROCEEDINGS{
+  author = {Travis Kessler, Eric R. Sacia, Alexis T. Bell, J. Hunter Mack}
+  title = {Predicting the Cetane Number of Furanic Biofuel Candidates Using an Improved Artificial Neural Network Based on Molecular Structure}
+  booktitle = {ASME 2016 Internal Combustion Engine Division Fall Technical Conference}
+  publisher = {ASME}
+  address = {Greenville, South Carolina}
+  year = {2016}
+  doi = {10.1115/ICEF2016-9383}
+}
+
+@article{
+  author = {Travis Kessler, Eric R. Sacia, Alexis T. Bell, J. Hunter Mack}
   title = {Artificial neural network based predictions of cetane number for furanic biofuel additives}
+  journal = {Fuel}
+  volume = {206}
+  pages = {171 - 179}
+  year = {2017}
   doi = {10.1016/j.fuel.2017.06.015}
-  howpublished = {\url{https://doi.org/10.1016/j.fuel.2017.06.015}}
 }
