### 2017-09-18 #401 @TJKessler
@arfon - all references in paper.md are now included in paper.bib.
