### 2017-09-20 #401 @arfon
Thanks @TJKessler - this is nearly right but not quite there. Please _remove_ the content below the `#References` section in your `paper.md` (https://github.com/TJKessler/ECNet/blob/master/paper/paper.md#references) and _cite_ the entries in your `paper.bib` as described [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html). That way, the citations are compiled into the `References` section when the PDF is created by Pandoc.
