+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@online{
+@online0{
   author = {Travis Kessler}
   title = {ECNet: Large scale machine learning projects for fuel property prediction}
   year = 2017
@@ -6,7 +6,7 @@
   urldate = {2017-10-04}
 }
 
-@online{
+@online1{
   author = {Travis Kessler}
   title = {ECNet 1.2.5.dev1}
   year = 2017
@@ -14,7 +14,7 @@
   urldate = {2017-10-04}
 }
 
-@online{
+@online2{
   author = {John Hunter Mack}
   title = {Energy and Combustion Research Laboratory}
   year = 2017
@@ -22,7 +22,7 @@
   urldate = {2017-10-04}
 }
 
-@PROCEEDINGS{
+@PROCEEDINGS0{
   author = {T. Sennott, C. Gotianun, R. Serres, M. Ziabasharhagh, J. H. Mack, R. Dibble}
   title = {Artificial Neural Network for Predicting Cetane Number of Biofuel Candidates Based on Molecular Structure}
   booktitle = {ASME 2013 Internal Combustion Engine Division Fall Technical Conference}
@@ -32,7 +32,7 @@
   doi = {10.1115/ICEF2013-19185}
 }
 
-@PROCEEDINGS{
+@PROCEEDINGS1{
   author = {Travis Kessler, Eric R. Sacia, Alexis T. Bell, J. Hunter Mack}
   title = {Predicting the Cetane Number of Furanic Biofuel Candidates Using an Improved Artificial Neural Network Based on Molecular Structure}
   booktitle = {ASME 2016 Internal Combustion Engine Division Fall Technical Conference}
@@ -42,7 +42,7 @@
   doi = {10.1115/ICEF2016-9383}
 }
 
-@article{
+@article0{
   author = {Travis Kessler, Eric R. Sacia, Alexis T. Bell, J. Hunter Mack}
   title = {Artificial neural network based predictions of cetane number for furanic biofuel additives}
   journal = {Fuel}
