### 2017-09-20 b4bc727
Update paper.bib
