### 2017-09-20 92be8ca
Update paper.md
