### 2017-09-21 b175135
Update paper.bib
