+++ paper/paper.bib
@@ -1,53 +1,53 @@
 @online{ecnet_github,
-  author = {Travis Kessler}
-  title = {ECNet: Large scale machine learning projects for fuel property prediction}
-  year = 2017
-  url = {https://github.com/TJKessler/ECNet}
+  author = {Travis Kessler},
+  title = {ECNet: Large scale machine learning projects for fuel property prediction},
+  year = {2017},
+  url = {https://github.com/TJKessler/ECNet},
   urldate = {2017-10-04}
 }
 
 @online{ecnet_pypi,
-  author = {Travis Kessler}
-  title = {ECNet 1.2.5.dev1}
-  year = 2017
-  url = {https://pypi.python.org/pypi/ecnet/}
+  author = {Travis Kessler},
+  title = {ECNet 1.2.5.dev1},
+  year = {2017},
+  url = {https://pypi.python.org/pypi/ecnet/},
   urldate = {2017-10-04}
 }
 
 @online{uml_ecrl,
-  author = {John Hunter Mack}
-  title = {Energy and Combustion Research Laboratory}
-  year = 2017
-  url = {http://faculty.uml.edu/Hunter_Mack/}
+  author = {John Hunter Mack},
+  title = {Energy and Combustion Research Laboratory},
+  year = {2017},
+  url = {http://faculty.uml.edu/Hunter_Mack/},
   urldate = {2017-10-04}
 }
 
 @PROCEEDINGS{asme_2013,
-  author = {T. Sennott, C. Gotianun, R. Serres, M. Ziabasharhagh, J. H. Mack, R. Dibble}
-  title = {Artificial Neural Network for Predicting Cetane Number of Biofuel Candidates Based on Molecular Structure}
-  booktitle = {ASME 2013 Internal Combustion Engine Division Fall Technical Conference}
-  publisher = {ASME}
-  address = {Dearborn, Michigan}
-  year = {2013}
+  author = {T. Sennott, C. Gotianun, R. Serres, M. Ziabasharhagh, J. H. Mack, R. Dibble},
+  title = {Artificial Neural Network for Predicting Cetane Number of Biofuel Candidates Based on Molecular Structure},
+  booktitle = {ASME 2013 Internal Combustion Engine Division Fall Technical Conference},
+  publisher = {ASME},
+  address = {Dearborn, Michigan},
+  year = {2013},
   doi = {10.1115/ICEF2013-19185}
 }
 
 @PROCEEDINGS{asme_2016,
-  author = {Travis Kessler, Eric R. Sacia, Alexis T. Bell, J. Hunter Mack}
-  title = {Predicting the Cetane Number of Furanic Biofuel Candidates Using an Improved Artificial Neural Network Based on Molecular Structure}
-  booktitle = {ASME 2016 Internal Combustion Engine Division Fall Technical Conference}
-  publisher = {ASME}
-  address = {Greenville, South Carolina}
-  year = {2016}
+  author = {Travis Kessler, Eric R. Sacia, Alexis T. Bell, J. Hunter Mack},
+  title = {Predicting the Cetane Number of Furanic Biofuel Candidates Using an Improved Artificial Neural Network Based on Molecular Structure},
+  booktitle = {ASME 2016 Internal Combustion Engine Division Fall Technical Conference},
+  publisher = {ASME},
+  address = {Greenville, South Carolina},
+  year = {2016},
   doi = {10.1115/ICEF2016-9383}
 }
 
 @article{fuel_2017,
-  author = {Travis Kessler, Eric R. Sacia, Alexis T. Bell, J. Hunter Mack}
-  title = {Artificial neural network based predictions of cetane number for furanic biofuel additives}
-  journal = {Fuel}
[…]
