### 2017-09-21 #401 @arfon
> @arfon - I think I got it this time, let me know if it is still off.

Nearly! Please merge https://github.com/TJKessler/ECNet/pull/7. Then, could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
