### 2017-09-21 #401 @TJKessler
@arfon - ECNet#7 has been merged, here is the updated DOI: 10.5281/zenodo.910866
