### 2017-09-21 #401 @arfon
@TJKessler - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00401 ⚡️ 🚀 :boom:
